+++ download.php
@@ -8,6 +8,7 @@ $autoloader = require 'vendor/autoload.php';
 $autoloader->addPsr4('Ksi\\', __DIR__.'/model');
 
 include 'config.php';
+include 'lib/helper.php';
 
 $config = include_once 'settings.php';
 $settings = $config['settings'];
@@ -28,16 +29,3 @@ if (isWorkingHour()){
 $allQuote = Ksi\QuoteBuilder::downloadQuote($t,$min);
 
 echo(date('Y-m-d H:i:s').' : '.implode(',', $allQuote))."\n";
-
-function isWorkingHour(){
-    $d = date('w');
-    if ( $d == 6 || $d == 7 ) {
-        return false;
-    }
-    $h = date("G");
-    if ( $h >= 9 && $h <=18) {
-        return true;
-    }
-
-    return false;
-}